refactor(imagePreview): add index file for simpler import
