+++ pro/src/new_components/ImagePreview/index.ts
@@ -1,0 +1,2 @@
+export { ImagePreview } from './ImagePreview'
+export { ImagePreviewsWrapper } from './ImagePreviewsWrapper'
